face_detection/requirements.txt
Click==7.0
dlib==19.17.0
face-recognition==1.2.3
face-recognition-models==0.3.0
numpy==1.17.0
pandas==0.25.0
Pillow==6.1.0
psycopg2-binary==2.8.3
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
